Move tests into ParallelCompiler test file and get error line correctly when compiling absolute paths
